LayoutTests/platform/wpe/imported/w3c/web-platform-tests/editing/run/undo-redo-expected.txt



PASS Waiting for load...
PASS insertParagraph at start of a paragraph
PASS insertParagraph at middle of a paragraph
PASS insertParagraph at end of a paragraph
PASS insertParagraph at start of a listitem
PASS insertParagraph at middle of a listitem
PASS insertParagraph at end of a listitem
PASS insertLineBreak at start of a paragraph
PASS insertLineBreak at middle of a paragraph
PASS insertLineBreak at end of a paragraph
PASS insertLineBreak at start of a listitem
PASS insertLineBreak at middle of a listitem
PASS insertLineBreak at end of a listitem
PASS delete at start of second paragraph
PASS forwarddelete at end of first paragraph
FAIL delete at start of second paragraph starting with an emoji promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'this.editingHost.ownerDocument')"
PASS forwarddelete at end of first paragraph ending with an emoji
FAIL insertParagraph at start of a paragraph - first undo assert_equals: insertParagraph at start of a paragraph - first undo: should restore the initial innerHTML value expected "<p>abcdef</p>" but got "<p><br></p><p>abcdef</p>"
FAIL insertParagraph at start of a paragraph - first redo assert_true: insertParagraph at start of a paragraph - first redo: execCommand("redo") should return true expected true got false
PASS insertParagraph at start of a paragraph - second redo
FAIL insertParagraph at middle of a paragraph - first undo assert_equals: insertParagraph at middle of a paragraph - first undo: should restore the initial innerHTML value expected "<p>abcdef</p>" but got "<p>abc</p><p>def</p>"
FAIL insertParagraph at middle of a paragraph - first redo assert_true: insertParagraph at middle of a paragraph - first redo: execCommand("redo") should return true expected true got false
PASS insertParagraph at middle of a paragraph - second redo
FAIL insertParagraph at end of a paragraph - first undo assert_equals: insertParagraph at end of a paragraph - first undo: should restore the initial innerHTML value expected "<p>abcdef</p>" but got "<p>abcdef</p><p><br></p>"
FAIL insertParagraph at end of a paragraph - first redo assert_true: insertParagraph at end of a paragraph - first redo: execCommand("redo") should return true expected true got false
PASS insertParagraph at end of a paragraph - second redo
FAIL insertParagraph at start of a listitem - first undo assert_equals: insertParagraph at start of a listitem - first undo: should restore the initial innerHTML value expected "<ul><li>abcdef</li></ul>" but got "<ul><li><br></li><li>abcdef</li></ul>"
FAIL insertParagraph at start of a listitem - first redo assert_true: insertParagraph at start of a listitem - first redo: execCommand("redo") should return true expected true got false
PASS insertParagraph at start of a listitem - second redo
FAIL insertParagraph at middle of a listitem - first undo assert_equals: insertParagraph at middle of a listitem - first undo: should restore the initial innerHTML value expected "<ul><li>abcdef</li></ul>" but got "<ul><li>abc</li><li>def</li></ul>"
FAIL insertParagraph at middle of a listitem - first redo assert_true: insertParagraph at middle of a listitem - first redo: execCommand("redo") should return true expected true got false
PASS insertParagraph at middle of a listitem - second redo
FAIL insertParagraph at end of a listitem - first undo assert_equals: insertParagraph at end of a listitem - first undo: should restore the initial innerHTML value expected "<ul><li>abcdef</li></ul>" but got "<ul><li>abcdef</li><li><br></li></ul>"
FAIL insertParagraph at end of a listitem - first redo assert_true: insertParagraph at end of a listitem - first redo: execCommand("redo") should return true expected true got false
PASS insertParagraph at end of a listitem - second redo
FAIL insertLineBreak at start of a paragraph - first undo assert_equals: insertLineBreak at start of a paragraph - first undo: should restore the initial innerHTML value expected "<p>abcdef</p>" but got "<p><br>abcdef</p>"
FAIL insertLineBreak at start of a paragraph - first redo assert_true: insertLineBreak at start of a paragraph - first redo: execCommand("redo") should return true expected true got false
PASS insertLineBreak at start of a paragraph - second redo
FAIL insertLineBreak at middle of a paragraph - first undo assert_equals: insertLineBreak at middle of a paragraph - first undo: should restore the initial innerHTML value expected "<p>abcdef</p>" but got "<p>abc<br>def</p>"
FAIL insertLineBreak at middle of a paragraph - first redo assert_true: insertLineBreak at middle of a paragraph - first redo: execCommand("redo") should return true expected true got false
PASS insertLineBreak at middle of a paragraph - second redo
FAIL insertLineBreak at end of a paragraph - first undo assert_equals: insertLineBreak at end of a paragraph - first undo: should restore the initial innerHTML value expected "<p>abcdef</p>" but got "<p>abcdef<br><br></p>"
FAIL insertLineBreak at end of a paragraph - first redo assert_true: insertLineBreak at end of a paragraph - first redo: execCommand("redo") should return true expected true got false
PASS insertLineBreak at end of a paragraph - second redo
FAIL insertLineBreak at start of a listitem - first undo assert_equals: insertLineBreak at start of a listitem - first undo: should restore the initial innerHTML value expected "<ul><li>abcdef</li></ul>" but got "<ul><li><br>abcdef</li></ul>"
FAIL insertLineBreak at start of a listitem - first redo assert_true: insertLineBreak at start of a listitem - first redo: execCommand("redo") should return true expected true got false
PASS insertLineBreak at start of a listitem - second redo
FAIL insertLineBreak at middle of a listitem - first undo assert_equals: insertLineBreak at middle of a listitem - first undo: should restore the initial innerHTML value expected "<ul><li>abcdef</li></ul>" but got "<ul><li>abc<br>def</li></ul>"
FAIL insertLineBreak at middle of a listitem - first redo assert_true: insertLineBreak at middle of a listitem - first redo: execCommand("redo") should return true expected true got false
PASS insertLineBreak at middle of a listitem - second redo
FAIL insertLineBreak at end of a listitem - first undo assert_equals: insertLineBreak at end of a listitem - first undo: should restore the initial innerHTML value expected "<ul><li>abcdef</li></ul>" but got "<ul><li>abcdef<br><br></li></ul>"
FAIL insertLineBreak at end of a listitem - first redo assert_true: insertLineBreak at end of a listitem - first redo: execCommand("redo") should return true expected true got false
PASS insertLineBreak at end of a listitem - second redo
FAIL delete at start of second paragraph - first undo assert_equals: delete at start of second paragraph - first undo: should restore the initial innerHTML value expected "<p>abc</p><p>def</p>" but got "<p>abcdef</p>"
FAIL delete at start of second paragraph - first redo assert_true: delete at start of second paragraph - first redo: execCommand("redo") should return true expected true got false
PASS delete at start of second paragraph - second redo
FAIL forwarddelete at end of first paragraph - first undo assert_equals: forwarddelete at end of first paragraph - first undo: should restore the initial innerHTML value expected "<p>abc</p><p>def</p>" but got "<p>abcdef</p>"
FAIL forwarddelete at end of first paragraph - first redo assert_true: forwarddelete at end of first paragraph - first redo: execCommand("redo") should return true expected true got false
PASS forwarddelete at end of first paragraph - second redo
FAIL forwarddelete at end of first paragraph ending with an emoji - first undo assert_equals: forwarddelete at end of first paragraph ending with an emoji - first undo: should restore the initial innerHTML value expected "<p>abc👉</p><p>👈def</p>" but got "<p>abc👉👈def</p>"
FAIL forwarddelete at end of first paragraph ending with an emoji - first redo assert_true: forwarddelete at end of first paragraph ending with an emoji - first redo: execCommand("redo") should return true expected true got false
PASS forwarddelete at end of first paragraph ending with an emoji - second redo
